src/agent/requirements.txt
anthropic>=0.42.0
google-genai>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
